use xml marshaling to render error
